+++ tests/apis/prv_api/test_providers.py
@@ -9,19 +9,20 @@ from tests.auth_helpers import auth_header, BASE_NUM_QUERIES
 @pytest.mark.django_db
 def test_provider_basic(client, django_assert_num_queries):
     provider = factories.Provider(
-        id=uuid.UUID("aaaaaaa0-1342-413b-8e89-db802b2f83f6"),
-        name="Test provider",
-        device_category="car",
+        id=uuid.UUID("aaaaaaa0-1342-413b-8e89-db802b2f83f6"), name="Test provider"
     )
-    factories.Provider(
-        id=uuid.UUID("bbbbbbb0-1342-413b-8e89-db802b2f83f6"),
-        name="Test provider 2",
-        device_category="scooter",
+    factories.Device(provider=provider, category="car")
+    factories.Device(
+        provider=factories.Provider(
+            id=uuid.UUID("bbbbbbb0-1342-413b-8e89-db802b2f83f6"), name="Test provider 2"
+        ),
+        category="scooter",
     )
-    factories.Provider(
-        id=uuid.UUID("ccccccc0-1342-413b-8e89-db802b2f83f6"),
-        name="Test provider 3",
-        device_category="bicycle",
+    factories.Device(
+        provider=factories.Provider(
+            id=uuid.UUID("ccccccc0-1342-413b-8e89-db802b2f83f6"), name="Test provider 3"
+        ),
+        category="bicycle",
     )
 
     response = client.get("/prv/providers/")
@@ -42,34 +43,37 @@ def test_provider_basic(client, django_assert_num_queries):
         "id": "aaaaaaa0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider",
         "logo_b64": None,
-        "device_category": "car",
+        "device_category": "",
         "base_api_url": "http://provider",
         "api_configuration": {"trailing_slash": False},
         "api_authentication": {"type": "none"},
         "agency_api_authentication": {"type": "none"},
         "colors": {},
+        "device_categories": {"bicycle": 0, "scooter": 0, "car": 1},
     } in response.data
     assert {
         "id": "bbbbbbb0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider 2",
-        "device_category": "scooter",
+        "device_category": "",
         "logo_b64": None,
         "base_api_url": "http://provider",
         "api_configuration": {"trailing_slash": False},
         "api_authentication": {"type": "none"},
         "agency_api_authentication": {"type": "none"},
         "colors": {},
+        "device_categories": {"bicycle": 0, "scooter": 1, "car": 0},
     } in response.data
     assert {
         "id": "ccccccc0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider 3",
         "logo_b64": None,
-        "device_category": "bicycle",
+        "device_category": "",
         "base_api_url": "http://provider",
         "api_configuration": {"trailing_slash": False},
         "api_authentication": {"type": "none"},
         "agency_api_authentication": {"type": "none"},
         "colors": {},
+        "device_categories": {"bicycle": 1, "scooter": 0, "car": 0},
     } in response.data
 
     response = client.get(
@@ -80,11 +84,12 @@ def test_provider_basic(client, django_assert_num_queries):
     assert response.data == {
         "id": "aaaaaaa0-1342-413b-8e89-db802b2f83f6",
         "name": "Test provider",
-        "device_category": "car",
+        "device_category": "",
         "logo_b64": None,
         "base_api_url": "http://provider",
         "api_configuration": {"trailing_slash": False},
         "api_authentication": {"type": "none"},
         "agency_api_authentication": {"type": "none"},
         "colors": {},
+        "device_categories": {"bicycle": 0, "scooter": 0, "car": 1},
     }